+++ src/main/java/eu/tib/controller/DataciteCommonsController.java
@@ -1,6 +1,6 @@
 package eu.tib.controller;
 
-import eu.tib.controller.validation.Validator;
+import eu.tib.controller.validation.InputValidator;
 import eu.tib.service.ResponseService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiParam;
@@ -25,9 +25,27 @@ public class DataciteCommonsController {
     @Autowired
     private ResponseService responseService;
 
+    @GetMapping(value = "/organization", produces = "application/json")
+    public ResponseEntity<String> getOrganization(
+            @Valid @Pattern(regexp = InputValidator.ror)
+            @ApiParam("Complete ROR URL consisting of https://ror.org/ plus id")
+            @RequestParam String ror) {
+
+        final String id = "sparqlg/datacitecommons/organization";
+        log.info("Incoming Request for " + id + " with ror: " + ror);
+        StopWatch stopWatch = new StopWatch(id);
+        stopWatch.start(id);
+
+        ResponseEntity result = responseService.buildResponse(id, Collections.singletonMap("ror", ror));
+
+        stopWatch.stop();
+        log.info(id + " took " + stopWatch.getTotalTimeSeconds() + "s");
+        return result;
+    }
+
     @GetMapping(value = "/organizationPlusPersons", produces = "application/json")
     public ResponseEntity<String> getOrganizationPlusPersons(
-            @Valid @Pattern(regexp = Validator.ror)
+            @Valid @Pattern(regexp = InputValidator.ror)
             @ApiParam("Complete ROR URL consisting of https://ror.org/ plus id")
             @RequestParam String ror) {
 
@@ -43,9 +61,45 @@ public class DataciteCommonsController {
         return result;
     }
 
+    @GetMapping(value = "/organizationPlusPersonsPlusPublications", produces = "application/json")
+    public ResponseEntity<String> getOrganizationPlusPersonsPlusPublications(
+            @Valid @Pattern(regexp = InputValidator.ror)
+            @ApiParam("Complete ROR URL consisting of https://ror.org/ plus id")
+            @RequestParam String ror) {
+
+        final String id = "sparqlg/datacitecommons/orga2person2publication";
+        log.info("Incoming Request for " + id + " with ror: " + ror);
+        StopWatch stopWatch = new StopWatch(id);
+        stopWatch.start(id);
+
+        ResponseEntity result = responseService.buildResponse(id, Collections.singletonMap("ror", ror));
+
+        stopWatch.stop();
+        log.info(id + " took " + stopWatch.getTotalTimeSeconds() + "s");
+        return result;
+    }
+
+    @GetMapping(value = "/person", produces = "application/json")
+    public ResponseEntity<String> getPerson(
+            @Valid @Pattern(regexp = InputValidator.orcid)
+            @ApiParam("Complete Orcid URL consisting of https://orcid.org/ plus id")
+            @RequestParam String orcid) {
+
+        final String id = "sparqlg/datacitecommons/person";
+        log.info("Incoming Request for " + id + " with orcid: " + orcid);
+        StopWatch stopWatch = new StopWatch(id);
+        stopWatch.start(id);
+
+        ResponseEntity result = responseService.buildResponse(id, Collections.singletonMap("orcid", orcid));
+
+        stopWatch.stop();
+        log.info(id + " took " + stopWatch.getTotalTimeSeconds() + "s");
+        return result;
+    }
+
     @GetMapping(value = "/personPlusPublications", produces = "application/json")
     public ResponseEntity<String> getPersonPlusPublications(
-            @Valid @Pattern(regexp = Validator.orcid)
+            @Valid @Pattern(regexp = InputValidator.orcid)
             @ApiParam("Complete Orcid URL consisting of https://orcid.org/ plus id")
             @RequestParam String orcid) {
 
@@ -55,6 +109,24 @@ public class DataciteCommonsController {
         stopWatch.start(id);
 
         ResponseEntity result = responseService.buildResponse(id, Collections.singletonMap("orcid", orcid));
+
+        stopWatch.stop();
+        log.info(id + " took " + stopWatch.getTotalTimeSeconds() + "s");
+        return result;
+    }
+
+    @GetMapping(value = "/publications", produces = "application/json")
+    public ResponseEntity<String> getPublications(
+            @Valid @Pattern(regexp = InputValidator.doi)
+            @ApiParam("DOI of the publication")
+            @RequestParam String doi) {
+
+        final String id = "sparqlg/datacitecommons/work";
+        log.info("Incoming Request for " + id + " with doi: " + doi);
+        StopWatch stopWatch = new StopWatch(id);
+        stopWatch.start(id);
+
+        ResponseEntity result = responseService.buildResponse(id, Collections.singletonMap("doi", doi));
 
         stopWatch.stop();
         log.info(id + " took " + stopWatch.getTotalTimeSeconds() + "s");